+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Le format est basé sur [Tenez un Changelog](https://keepachangelog.com/fr/1.0.0
 ## [À venir dans la prochaine version]
 ### Ajouté
 - Bouton pour afficher / cacher un mot de posse
+### Modifié
+- Offre d'essais = Offre premium avec 3 mois
+